backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
chromadb==0.5.23
openai==1.58.1
rank-bm25==0.2.2
python-dotenv==1.0.1
python-multipart==0.0.18
PyPDF2==3.0.1
tiktoken==0.8.0
pydantic==2.10.3
numpy>=1.24.0
